* added gui_input_box (fully functional, with unicode text input, basic control key support (arrow keys, ctrl/cmd+*), mouse input/handling and blink character)  * added active gui object tracking to the gui (setter/getter); when gui objects become (in)active (or are set (in)active externally), they will call these gui functions; among other things, this is used to correctly activate/deactive input boxes  * instead of always fully locking the gui when drawing the gui windows, only try_lock, so rendering (and the engine) doesn't halt/freeze when the gui lock is held in another thread; this also prevents dead-locking when another thread needs both the gui lock and the engine lock  * enable/set scissor text for ui object rendering in gui_theme and added clear parameter (if set to false, rtt clear will not be called)  * fixed/actually enable fullscreen anti-aliased buffer sharing (color and depth attachment must be reset every time; set the correct depth type; use correct anti-aliasing mode in gui)  * fixed gfx2d border rendering of rectangles (weirdly, this is more complex than computing the border triangle strip of rounded rectangles)  * refactored the font code: added a text_stepper function (used to be create_text_ubo_data) which will step over a text, apply formatting and call user/externally specified functions for each character/code (this way, many different things can be done very easily with the same code, e.g. text caching/ubo data creation, computing the advance of a text, ...)  * the font cache function will now check if any characters/codes need to be cached at the start of the function (no need to change gl state if there is nothing to do); added is_cached function  * EVENT_TYPE: added __GUI_EVENT group, added CLIPBOARD_UPDATE event and did some cleanup (use bit-shifting to initialize __*_EVENT enums)  * GUI_EVENT enum class is now a subset of EVENT_TYPE (or: GUI_EVENT overlaps EVENT_TYPE); range starts at __GUI_EVENT+1  * added general support for gui event objects (added gui_event_base and gui_event)  * create gui_events for all calls to gui_object::handle  * added state getters to gui_object (+setter for state.active)  * correct parent<->child handling in gui_object  * delete all childs in the gui_window destructor (for convenience, so they don't need to be deleted separately)  * added clipboard support to the gui  * added thickness parameter/attribute to gui_theme lines  * more rtt cleanup (removed old code, correct state changes)

+++ src/rendering/rtt.h
@@ -82,7 +82,7 @@ public:
 		DEPTH_TYPE depth_type = DEPTH_TYPE::NONE;
 		STENCIL_TYPE stencil_type = STENCIL_TYPE::NONE;
 		GLenum depth_attachment_type = GL_DEPTH_ATTACHMENT;
-		size_t samples = 0;
+		unsigned int samples = 0;
 		
 		vector<unsigned int> tex;
 		vector<unsigned int> resolve_buffer;
